Switch default project build to scala 2.12
- geomesa does not cross publish for 2.12
- geowave does not corss publish for 2.12

These projects are both experimental so they're bing unlinked from the build until we either update or figure out the best way to deal with them.
