+++ version.sbt
@@ -1,0 +1,1 @@
+ThisBuild / version := "3.1.1-SNAPSHOT"
